requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
Flask-Login==0.5.0
Flask-WTF==0.15.1
Flask-Migrate==3.1.0
Flask-Caching==2.0.1
Flask-Limiter==2.6.3
Flask-Talisman==1.0.0
Flask-CORS==3.0.10
Flask-Session==0.4.0
APScheduler==3.9.1
beautifulsoup4==4.9.3
requests==2.26.0
python-dotenv==0.19.0
gunicorn==20.1.0
Werkzeug==2.0.1
SQLAlchemy==1.4.23
WTForms==2.3.3
email-validator==1.1.3
pytest==6.2.5
coverage==6.0.2
bleach==4.1.0
redis==4.1.0
python-jose==3.3.0
cryptography==37.0.2
psycopg2-binary==2.9.1
python-dateutil==2.8.2
pytz==2021.1
Pillow==8.3.2
markdown==3.3.4
